% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch-xcs.R
\name{calc_xcs}
\alias{calc_xcs}
\title{utility function to fetch xCS data for teams}
\usage{
calc_xcs(match_id, verbose = TRUE, details = FALSE)
}
\description{
utility function to fetch xCS data for teams
}
