% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_fixtures.R
\name{fetch_fixtures}
\alias{fetch_fixtures}
\title{Fetch Fixtures for next n gameweeks}
\usage{
fetch_fixtures(n)
}
\arguments{
\item{n}{number of fixtures to fetch}
}
\value{
a data.frame of fixtures.
}
\description{
gets fixtures for the next n gameweeks and optionally plots the difficulty.
}
\examples{
fixtures <- fetch_fixtures(n = 5)

# visualise fixtures
plot(fixtures)

# find teams with low correlation fixures (good for rotation)
fixture_rotation(fixtures)


}
