% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_understat_data.R
\name{fetch_understat_data}
\alias{fetch_understat_data}
\title{experimental script to fetch data from understat}
\usage{
fetch_understat_data(team = "Arsenal")
}
\description{
experimental script to fetch data from understat
}
