% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixture_rotation.R
\name{fixture_rotation}
\alias{fixture_rotation}
\alias{fixture_rotation.default}
\alias{fixture_rotation.fpl_fixtures}
\title{Fixture rotation of fpl_fixtures}
\usage{
fixture_rotation(x, ...)

\method{fixture_rotation}{default}(x, ...)

\method{fixture_rotation}{fpl_fixtures}(x, next_n = 10)
}
\arguments{
\item{x}{An object of class \code{fpl_fixtures}}
}
\value{
An object of class \code{fpl_fixtures}
}
\description{
Calculates correlations between all team's upcoming fixtures to give best rotation possibilities
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method not implemented yet.

\item \code{fpl_fixtures}: Lowest correlation fixtures for the next n games which should be good for sub rotation.
}}

\examples{
fixtures <- fetch_fixtures(n = 5)

# visualise fixtures
plot(fixtures)

# find teams with low correlation fixures (good for rotation)
fixture_rotation(fixtures)
}
