% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_team.R
\name{optimise_team}
\alias{optimise_team}
\title{Fetch Player Data From The FPL API}
\usage{
optimise_team(objective = "points", bank = 1000, bench_value = 170,
  gk = 1, def = 3, mid = 4, fwd = 3, gameweek_range = F,
  min_games = 1, custom_df = F, expected_points_adjust = F)
}
\arguments{
\item{objective}{takes value 'points', 'ppg' or 'vapm', indicating whether we want to maximise the total points obtained over the course of the dataset (season), the points per game played, or value added per million.}

\item{bank}{takes a numeric value indicating the budget available for our full team (team + bench). Enter the value in four digits so 100.0m would be 1000 - the default for the season start.}

\item{bench_value}{before running the optimiser you'll want to have chosen your bench already (there is little point in optimising players on your bench as they will not earn you points a lot of the time). Enter the value of your chosen bench in four digits as per the bank parameter.}

\item{gk}{number of goalkeepers we want to pick - most likely one. If rotating keepers, it can be useful to pick your goalkeepers and add both to your bench value, then set gk = 0. This prevents the goalkeeper being optimised as if it were in goal as set and forget.}

\item{def}{number of defenders to pick in our optimal solution}

\item{mid}{number of midfielders to pick in our optimal solution}

\item{fwd}{number of forwards to pick in our optimal solution}

\item{gameweek_range}{numeric vector of the gameweeks to include in the data. This is not available pre-season and defaults to FALSE including all historic data for the past season.}

\item{min_games}{the minimum number of games a player must have played to be considered for selection. This should be set when the objective function is 'ppg' so that players with high points per game but with relatively few games played are avoided.}

\item{custom_df}{defaults to FALSE. If you want to make adjustments to the player data before optimising (eg. you want to exclude certain players from consideration) you can use the fetch_player_data() function and update the resulting dataset yourself. This dataset can be passed as the custom_df parameter.}

\item{expected_points_adjust}{defaults to FALSE. Should we adjust points and points per game for the xA / xG? Data comes from understat.}
}
\value{
a data.frame of optimised solution.
}
\description{
Solves the optimisation problem for a FPL team given a budget constraint, amount of players, and objective function.
}
\examples{
df <- fetch_player_data()

# RUN OPTIMISATION  ON POINTS PER GAME FOR A 4-3-3 formation, with a bench value of 175 and minimum games of 13
result <- optimise_team(objective = 'ppg', custom_df = F, bench_value = 175, gk = 1, def = 4, mid = 3, fwd = 3, min_games = 13)
sum(result$points_per_game)


# RUN OPTIMISATION  ON TOTAL POINTS FOR A 3-4-3 formation, with a bench value of 170
result <- optimise_team(objective = 'points', custom_df = F, bench_value = 170, gk = 1, def = 3, mid = 4, fwd = 3)
sum(result$total_points)

# USING CUSTOM DATA TO FILTER OUT UNWANTED RESULTS
df <- fetch_player_data() \%>\%
  # must have played more than 13 games
  filter(games >= 13) \%>\%
  # I don't want to consider Milivojevic or Alonso
  filter(! id \%in\% c(134, 103))

# adjust zaha's points as he's now midfield (1 extra pt for goal and point for cs)
df[df$id == 133,]$total_points <- df[df$id == 133,]$total_points + df[df$id == 133,]$goals_scored + df[df$id == 133,]$clean_sheets
df[df$id == 133,]$points_per_game <- df[df$id == 133,]$total_points / df[df$id == 133,]$games

# same for perez
df[df$id == 265,]$total_points <- df[df$id == 265,]$total_points + df[df$id == 265,]$goals_scored + df[df$id == 265,]$clean_sheets
df[df$id == 265,]$points_per_game <- df[df$id == 265,]$total_points / df[df$id == 265,]$games

# downgrade jota and delofeou's stats
df[df$id == 363,]$total_points <- df[df$id == 363,]$total_points - df[df$id == 363,]$goals_scored - df[df$id == 363,]$clean_sheets
df[df$id == 363,]$points_per_game <- df[df$id == 363,]$total_points / df[df$id == 363,]$games

df[df$id == 410,]$total_points <- df[df$id == 410,]$total_points - df[df$id == 410,]$goals_scored - df[df$id == 410,]$clean_sheets
df[df$id == 410,]$points_per_game <- df[df$id == 410,]$total_points / df[df$id == 410,]$games

# RUN OPTIMISATION WITH CUSTOM DATA (using custom_df = df option)
result <- optimise_team(objective = 'ppg', custom_df = df, bench_value = 175, gk = 1, def = 4, mid = 4, fwd = 2, min_games = 14)
sum(result$points_per_game)


# RUN WITH expected goals / assists
result <- optimise_team(objective = 'ppg', custom_df = df, bench_value = 175, gk = 1, def = 4, mid = 4, fwd = 2, min_games = 14, expected_points_adjust = T)
sum(result$points_per_game)




}
