% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_player_data.R, R/fetch_xcs_ha.R
\name{\%>\%}
\alias{\%>\%}
\alias{fetch_player_data}
\alias{fetch_xCS_ha}
\title{Fetch Player Data From The FPL API}
\usage{
fetch_player_data(gameweek_range = 1:2, force_refresh = F)

fetch_xCS_ha(year = 2019, ...)
}
\arguments{
\item{gameweek_range}{numeric vector of the gameweeks to include in the data. This is not available pre-season and defaults to FALSE including all historic data for the past season.}

\item{year}{Season we want to download the xCS data for. For 2019/20 enter 2019.}
}
\value{
a data.frame of the full player data and history.

a data.frame of the expected vs actual clean sheets, split by home/away.
}
\description{
Fetches the historical data for each FPL player, including the current prices of those players.

Fetches the expected clean sheet data for each prem team. Uses shot data scraped from Understat.
}
\examples{
df <- fetch_player_data()

df <- fetch_player_data()

}
