% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpl_fixtures.R
\name{plot.fpl_fixtures}
\alias{plot.fpl_fixtures}
\title{Visualize Fixture Difficulty}
\usage{
\method{plot}{fpl_fixtures}(x)
}
\arguments{
\item{x}{An object of class \code{fpl_fixtures}.}
}
\value{
An object of class \code{ggplot2}.
}
\description{
GGplot visualization of an object of class \code{fpl_fixtures}.
}
\examples{
fixtures <- fetch_fixtures(n = 5)

# visualise fixtures
plot(fixtures)
}
\seealso{
\code{\link{fetch_fixtures}}.
}
