% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fetch-understat.R
\name{fetch_understat_data}
\alias{fetch_understat_data}
\title{utility function to fetch data for players of a given team from understat}
\usage{
fetch_understat_data(team = "Arsenal", year = "2020")
}
\description{
utility function to fetch data for players of a given team from understat
}
