% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_understat_data.R
\name{fetch_xg_data}
\alias{fetch_xg_data}
\title{Fetch xG and xA data from understat}
\usage{
fetch_xg_data()
}
\value{
a data.frame of the full player xg data for the season.
}
\description{
Scrapes understat data for the 2018 season from understat for the 17 teams that were in the PL in both 2018 and 2019 seasons
}
\examples{
df <- fetch_understat_data()


}
