% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_xg_data.R
\name{fetch_xg_data}
\alias{fetch_xg_data}
\title{Fetch xG and xA data from understat}
\usage{
fetch_xg_data(year = 2023, reduce = FALSE, check_data = FALSE)
}
\arguments{
\item{year}{the season for which we want to scrape data from understat.}

\item{reduce}{If TRUE, returns only select columns from the API for nicer display. If FALSE, will return all columns from the FPL API.}

\item{check_data}{used for debugging only}
}
\value{
a data.frame of the full player xg data for the season.
}
\description{
Scrapes xA and xG data for a given season from understat
}
\examples{
df <- fetch_xg_data(year = 2023)


}
