% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_player_data.R, R/fetch_xcs.R
\name{\%>\%}
\alias{\%>\%}
\alias{fetch_player_data}
\alias{fetch_xCS}
\title{Fetch Player Data From The FPL API}
\usage{
fetch_player_data(reduce = TRUE)

fetch_xCS(year = 2023, match_id = "all", ungroup = F, ...)
}
\arguments{
\item{reduce}{If TRUE, returns only select columns from the API for nicer display. If FALSE, will return all columns from the FPL API.}

\item{year}{Season we want to download the xCS data for. For 2020/21 enter 2021.}

\item{match_id}{Fetch xCS data for a single match, if you know the match ID. By default it is set to 'all'.}

\item{ungroup}{Set this to TRUE if you want raw data for each game.}
}
\value{
a data.frame of the full player data and history.

a data.frame of the expected clean sheets, split by home/away.
}
\description{
Fetches the historical data for each FPL player, including the current prices of those players.

Fetches the expected clean sheet data for each prem team. Uses shot data scraped from Understat.
}
\examples{
df <- fetch_player_data()

df <- fetch_xCS()

}
