% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_player_data.R
\name{\%>\%}
\alias{\%>\%}
\alias{fetch_player_data}
\title{Fetch Player Data From The FPL API}
\usage{
fetch_player_data(gameweek_range = F)
}
\arguments{
\item{gameweek_range}{numeric vector of the gameweeks to include in the data. This is not available pre-season and defaults to FALSE including all historic data for the past season.}
}
\value{
a data.frame of the full player data and history.
}
\description{
Fetches the historical data for each FPL player, including the current prices of those players.
}
\examples{
df <- fetch_player_data()

}
